move implementation details of PluginDiscoveryTask to separate class
