+++ src/test/groovy/org/shipkit/internal/gradle/plugin/PluginDiscoveryTest.groovy
@@ -1,0 +1,26 @@
+package org.shipkit.internal.gradle.plugin
+
+import testutil.PluginSpecification
+import static PluginDiscoveryPluginTest.META_INF_GRADLE_PLUGINS
+
+class PluginDiscoveryTest extends PluginSpecification {
+
+    def "generate plugin name"() {
+        expect:
+        PluginDiscovery.generatePluginName(input) == expected
+
+        where:
+        input                                       | expected
+        'plugin.properties'                         | 'plugin'
+        'com.shipkit.base-java-library.properties'  | 'baseJavaLibrary'
+        'com.shipkit.versioning.properties'         | 'versioning'
+    }
+
+    def "implementation class"() {
+        when:
+        project.file(META_INF_GRADLE_PLUGINS).mkdirs()
+        File file = project.file("$META_INF_GRADLE_PLUGINS/org.shipkit.plugin-name-sample.properties") << "implementation-class=org.shipkit.PluginNameSample"
+        then:
+        PluginDiscovery.getImplementationClass(file) == 'org.shipkit.PluginNameSample'
+    }
+}
